Change type from signed to unsigned to avoid negative sizes

+++ doublebeam-cpp/src/utils.hpp
@@ -86,7 +86,7 @@ namespace seismo {
      * is the x axis, the second axis is the y axis and the third axis contains the z values.
      */
     std::vector<position_t> grid_coordinates(double x0, double x1, double y0, double y1,
-                                             double depth, int num_x, int num_y);
+                                             double depth, size_t num_x, size_t num_y);
 } // namespace seismo
